Manifold expansion working for uMPS
There's a subtlety when starting from a
product state (xi = 1). If such state is
built randomly the TDVP will get stuck in
the manifold without being able to expand
to a larger xiTilde. This happens when the
quantum fluctuactions are quenched by the
xi = 1 manifold. On the other hand, if the
Hamiltonian is "more classical", in the
sense of having large degrees of freedom,
then the TDVP is capable of increasing the
size of the manifold and doing the
expansion.

For instance, for d = 2 spin systems, the
only way of to get the manifold to expand,
when starting with xi = 1, is by starting
with an infinite-temperature state. If, on
the contrary, d = 3 then any ramdom intial
product state will do and the dynamic
expansion happens smoothly.

That's why I've decided to define different
thresholds for xi == 1 and xi != 1. Other
than that, the dynamic expansion is done
with no jumps in the energy or the fidelity.