Thld changed; no truncation; xi dyn. path
It seems that getting the energy to high
precision depends quite a lot in the way
the bond dimension is dynamically
increased. If the bond dimension in
increased like xi(it+1) = xi(it) * d, then
after some "critical" value xi_c, we start
seeing zero Schmidt coefficients. It also
seems that the TDVP is not capable of
recovering from this "effective truncation"
by mixing the basis. This is expected to
happen since as the more converged the
state is the more diagonal the eigenvectors
of the tranfer matrix are. If we change the
path of the bond dimension to
xi(it+1) = xi(it) + const. * [d, xi(it)],
then it's less likely to encounter zero
eigenvalues since the Schmidt weights have
to be distributed over a smaller basis
compared to the former path. When we take
the smoother path then we can go to large
bond dimensions reaching accurate results.

This situation worsens for large bond
dimensions since the contribution of B01
and B10 to the new MPS are increasingly
smaller. In addition, if we cut down states
with a larger threshold then more zero
Schmidt coefficients there will be,
triggering the truncation or not helping
with the expansion of the manifold.

If this situation turns out to be true then
having the option of decreasing the bond
dimension will either be innocuous or fight
against the dynamic expansion. That's why
I've decided to remove the truncation
feature and diminish the value of the
threshold 'thld'.

Maybe this is an issue with the time step
or how accurate if the convergence
criterion for a given xi. If we decrease
the step or perhaps increase eta then we
can take longer leaps in the expansion of
the manifold.

In any case, with the code as is we get an
energy difference of ~10**(-11) with D~900,
for the S=1 chain. For the XY model (h=0)
we get ~10**(-7) with D~300, but of course
it takes much longer to get that precision.
These energies are compared to the
corresponding exact solutions.